+++ changelog.md
@@ -3,6 +3,7 @@
 ## 0.0.3
 - Modified ore generation to be slightly less plentiful in caves
 - Tweaked cave generation to create larger caverns more often and to generate vertical cracks less often
+- Fixed an issue where grass would grow underwater
 
 ## 0.0.2
 - Added translations for mod blocks and items